// src/types/NoodeClassNames.ts
export default interface NoodeClassNames {
    noode: string[];
    contentBox: string[];
    childIndicator: string[];
    overflowIndicatorLeft: string[];
    overflowIndicatorRight: string[];
    overflowIndicatorTop: string[];
    overflowIndicatorBottom: string[];
    branch: string[];
    branchBackdrop: string[];
}