// src/types/NoodeSerializedCss.ts
export default interface NoodeSerializedCss {
    noode?: string;
    contentBox?: string;
    childIndicator?: string;
    overflowIndicatorLeft?: string;
    overflowIndicatorRight?: string;
    overflowIndicatorTop?: string;
    overflowIndicatorBottom?: string;
    branch?: string;
    branchBackdrop?: string;
}