// src/types/NoodeStyles.ts
export default interface NoodeStyles {
    noode: object;
    contentBox: object;
    childIndicator: object;
    overflowIndicatorLeft: object;
    overflowIndicatorRight: object;
    overflowIndicatorTop: object;
    overflowIndicatorBottom: object;
    branch: object;
    branchBackdrop: object;
}